+++ es6/tests/e2e/v3.js
@@ -1,18 +1,48 @@
+const { cloneDeep } = require("lodash");
 const {
 	resolveSoon,
+	cleanRecursive,
 	createDocV3,
 	captureLogs,
-	makeDocx,
+	makeDocxV3,
+	createXmlTemplaterDocxNoRender,
 	expect,
 	shouldBeSame,
 	expectToThrow,
 	wrapMultiError,
+	expectToThrowAsync,
 } = require("../utils.js");
 const Errors = require("../../errors.js");
 const rawXMLValue = require("../data/raw-xml-pptx.js");
 
 describe("V3 API", function () {
+	beforeEach(function () {
+		this.deprecations = null;
+		this.capture = captureLogs();
+	});
+	afterEach(function () {
+		this.capture.stop();
+		if (this.deprecations != null) {
+			const logs = this.capture.logs();
+			this.deprecations.forEach(function (deprecation) {
+				const found = logs.some(function (log) {
+					if (log.indexOf(deprecation) !== -1) {
+						return true;
+					}
+				});
+				if (!found) {
+					throw new Error(`Expected to get deprecation '${deprecation}'`);
+				}
+			});
+		}
+	});
 	it("should work with setOptions", function () {
+		this.deprecations = [
+			"Deprecated docxtemplater constructor with no arguments, view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide",
+			'Deprecated method ".setOptions", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+			'Deprecated method ".attachModule", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+			'Deprecated method ".loadZip", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+		];
 		const doc = createDocV3("tag-multiline.docx");
 		expect(() => doc.setOptions()).to.throw(
 			"setOptions should be called with an object as first parameter"
@@ -20,17 +50,15 @@ describe("V3 API", function () {
 	});
 
 	it("should work when the delimiters are passed", function () {
-		const options = {
-			delimiters: {
-				start: "<",
-				end: ">",
-			},
-		};
+		this.deprecations = [
+			'Deprecated method ".setData", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+		];
+		const options = { delimiters: { start: "<", end: ">" } };
 		const doc = createDocV3("delimiter-gt.docx", options);
-		doc.compile();
-		doc.render({
+		doc.setData({
 			user: "John",
 		});
+		doc.render();
 		const fullText = doc.getFullText();
 		expect(fullText).to.be.equal("Hello John");
 	});
@@ -61,6 +89,11 @@ describe("V3 API", function () {
 	});
 
 	it("should work with simple raw pptx async", function () {
+		this.deprecations = [
+			'Deprecated method ".compile", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide, stack',
+			'Deprecated method ".resolveData", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+			'Deprecated method ".loadZip", view upgrade guide : https://docxtemplater.com/docs/api/#upgrade-guide',
+		];
 		let scope, meta, tag;
 		let calls = 0;
 		const doc = createDocV3("raw-xml-example.pptx", {
@@ -101,13 +134,194 @@ describe("V3 API", function () {
 				offset: 0,
 			},
 		};
-		const doc = makeDocx(content, { errorLogging: false });
-		const capture = captureLogs();
+
 		expectToThrow(
-			() => doc.compile(),
+			() => makeDocxV3(content, { errorLogging: false }).compile(),
 			Errors.XTTemplateError,
 			wrapMultiError(expectedError)
 		);
-		capture.stop();
+	});
+
+	it("should show error when running resolveData before compile", function () {
+		const content = "<w:t>{#users}{user}{/}</w:t>";
+		const expectedError = {
+			name: "InternalError",
+			message: "You must run `.compile()` before running `.resolveData()`",
+			properties: {
+				id: "resolve_before_compile",
+			},
+		};
+		const doc = makeDocxV3(content);
+		return expectToThrowAsync(
+			() => doc.resolveData(),
+			Errors.XTInternalError,
+			expectedError
+		);
+	});
+
+	it("should resolve with simple table", function () {
+		const doc = createDocV3("table-complex2-example.docx");
+		doc.compile();
+		return doc
+			.resolveData({
+				table1: [
+					{
+						t1data1: "t1-1row-data1",
+						t1data2: "t1-1row-data2",
+						t1data3: "t1-1row-data3",
+						t1data4: "t1-1row-data4",
+					},
+					{
+						t1data1: "t1-2row-data1",
+						t1data2: "t1-2row-data2",
+						t1data3: "t1-2row-data3",
+						t1data4: "t1-2row-data4",
+					},
+					{
+						t1data1: "t1-3row-data1",
+						t1data2: "t1-3row-data2",
+						t1data3: "t1-3row-data3",
+						t1data4: "t1-3row-data4",
+					},
+				],
+				t1total1: "t1total1-data",
+				t1total2: "t1total2-data",
+				t1total3: "t1total3-data",
+			})
+			.then(function (resolved) {
+				const myresolved = cloneDeep(resolved);
+				cleanRecursive(myresolved);
+				expect(myresolved).to.be.deep.equal([
+					{
+						tag: "t1total1",
+						value: "t1total1-data",
+					},
+					{
+						tag: "t1total2",
+						value: "t1total2-data",
+					},
+					{
+						tag: "t1total3",
+						value: "t1total3-data",
+					},
+					{
+						tag: "table1",
+						value: [
+							[
+								{
+									tag: "t1data1",
+									value: "t1-1row-data1",
+								},
+								{
+									tag: "t1data2",
+									value: "t1-1row-data2",
+								},
+								{
+									tag: "t1data3",
+									value: "t1-1row-data3",
+								},
+								{
+									tag: "t1data4",
+									value: "t1-1row-data4",
+								},
+							],
+							[
+								{
+									tag: "t1data1",
+									value: "t1-2row-data1",
+								},
+								{
+									tag: "t1data2",
+									value: "t1-2row-data2",
+								},
+								{
+									tag: "t1data3",
+									value: "t1-2row-data3",
+								},
+								{
+									tag: "t1data4",
+									value: "t1-2row-data4",
+								},
+							],
+							[
+								{
+									tag: "t1data1",
+									value: "t1-3row-data1",
+								},
+								{
+									tag: "t1data2",
+									value: "t1-3row-data2",
+								},
+								{
+									tag: "t1data3",
+									value: "t1-3row-data3",
+								},
+								{
+									tag: "t1data4",
+									value: "t1-3row-data4",
+								},
+							],
+						],
+					},
+				]);
+				doc.render();
+				const fullText = doc.getFullText();
+				expect(fullText).to.be.equal(
+					"TABLE1COLUMN1COLUMN2COLUMN3COLUMN4t1-1row-data1t1-1row-data2t1-1row-data3t1-1row-data4t1-2row-data1t1-2row-data2t1-2row-data3t1-2row-data4t1-3row-data1t1-3row-data2t1-3row-data3t1-3row-data4TOTALt1total1-datat1total2-datat1total3-data"
+				);
+			});
+	});
+
+	describe("Using the resolveOffset property", function () {
+		it("should work", function () {
+			const content =
+				"<w:t>{#loop_first}{#loop_second}{name_inner} {name_outer}{/loop_second}{/loop_first}</w:t>";
+			const xmlt = createXmlTemplaterDocxNoRender(content, {}).compile();
+			return xmlt
+				.resolveData({
+					loop_first: [1],
+					loop_second: [{ name_inner: "John" }],
+					name_outer: "Henry",
+				})
+				.then(function () {
+					const sm = xmlt.scopeManagers["word/document.xml"];
+
+					expect(sm.finishedResolving).to.equal(true);
+
+					sm.scopePath.unshift("aaa");
+					sm.scopePathItem.unshift(122);
+					sm.scopePathLength.unshift(144);
+					sm.scopeLindex.unshift(555);
+					sm.resolveOffset = 1;
+
+					const part = {
+						value: "loop_first",
+						lIndex: 3,
+					};
+					const part2 = {
+						value: "loop_second",
+						lIndex: 6,
+					};
+					let val;
+
+					function loopOver(scope, i, length) {
+						const ssm = sm.createSubScopeManager(
+							scope,
+							part.value,
+							i,
+							part,
+							length
+						);
+
+						val = ssm.getValue("loop_second", { part: part2 });
+					}
+
+					sm.loopOver("loop_first", loopOver, false, {
+						part: { value: "loop_first", lIndex: 3 },
+					});
+					expect(val[0][0].tag).to.equal("name_inner");
+					expect(val[0][0].value).to.equal("John");
+				});
+		});
 	});
 });